+++ Assets/Item/Script/HealthItem/Interface/IHealthFlask.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IHealthFlask
 {
-   public int HealthValue { get; }
+   public float HealthValue { get; }
 }
