+++ Assets/Item/Script/MagicItem/Interface/IManaFlask.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IManaFlask 
 {
-    public int ManaValue { get; }
+    public float ManaValue { get; }
 }
