+++ Assets/Player/ScriptebleObject/Script/CharacterMagicSO.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "ScriptableObject/Character/Mana",fileName = "newManaPlayer")]
 public class CharacterManaSO : ScriptableObject
 {
-    public int manaCharacter;
+    public float manaCharacter;
 }
